Prevent memory allocaiton and pointer problems
The bigest problem was in process_response where the pointer
last_response was pointing at an argument of the function which is a
temporary variable. So the pointer was not valid anymore outside of
process_response.

Furthermore, by making last_response and result regular members, the
user of this class is no longer responsible for deleting result,
resulting lower risk of memory leaks.