Make sure the command id is unique within MqttRpcclientQt instance
Before this commit, the command id was generated by OpCommand, but it
was never checked if the command was in fact unique (which it needs to
be because it is later placed in a map and for identification). Now
MqttRpcclientQt creates a guaranteed (locally) unique commandId and
passes that to OpCommand when serializing the command.