mqtt_rpc_client_qt: Cleanup before disconnection
In some situations, a crash may occur when disconnecting
from the host specially when it is heavily used prior the
disconnection, for example during the first seconds after
connecting to the broker.
Disconnection from hostis also triggered during a delete.
Using deleteLater() doesn't improve the situation.

To prevent it, you must perform cleanup before calling
disconnectFromHost() and wait until state changes to
disconnected before deleting the object.
The cleanup includes unsubscribing from all topics and
waiting until it is completed.

The closeAndDelete() method was added for this purpose and it
should be used instead of delete or deleteLater().